+++ src/elements/formes/Chemin.cpp
@@ -11,10 +11,10 @@ std::string Chemin::to_svg() const {
 
 	std::for_each(++_points.begin(), _points.end(), [&result, &precedent, &proprietes] (Point const& p) {
 		result += std::string("<line ")
-			+ "x1=\"" + std::to_string(precedent.x) + "\" "
-			+ "y1=\"" + std::to_string(precedent.y) + "\" "
-			+ "x2=\"" + std::to_string(p.x) + "\" "
-			+ "y2=\"" + std::to_string(p.y) + "\" "
+			+ "x1=\"" + std::to_string(precedent.x.toDouble()) + "\" "
+			+ "y1=\"" + std::to_string(precedent.y.toDouble()) + "\" "
+			+ "x2=\"" + std::to_string(p.x.toDouble()) + "\" "
+			+ "y2=\"" + std::to_string(p.y.toDouble()) + "\" "
 			+ proprietes
 			+ "/>";
 		precedent = p;
@@ -31,7 +31,7 @@ Forme::Point Chemin::centre() const {
 		result.y += p.y;
 	});
 
-	return { result.x / _points.size(), result.y / _points.size() };
+	return { result.x.toDouble() / _points.size(), result.y.toDouble() / _points.size() };
 }
 
 void Chemin::ajoutePoint(double x, double y) {
